+++ CHANGELOG.md
@@ -17,6 +17,7 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 - Visibility of method `HarParser::getHttpRequest(Request $request): RequestInterface` to public
 - Visibility of method `HarParser::getHttpResponse(Response $response): ResponseInterface` to public
 - Merge options of request with global options
+- Use previous request URI if "baseUri" option is not defined
 
 ### Fixed